IGNITE-24831 Move MessageCollectionItemType to network-api
